Distribution.CabalSpecVersion

Documentation

data CabalSpecVersion

Different Cabal-the-spec versions.

We branch based on this at least in the parser.

Constructors

CabalSpecV1_0
this is older than CabalSpecV1_2

CabalSpecV1_2
new syntax (sections)

CabalSpecV1_4

CabalSpecV1_6

CabalSpecV1_8

CabalSpecV1_10

CabalSpecV1_12

CabalSpecV1_18

CabalSpecV1_20

CabalSpecV1_22

CabalSpecV1_24

CabalSpecV2_0

CabalSpecV2_2

CabalSpecV2_4

CabalSpecV3_0

CabalSpecV3_4

CabalSpecV3_6

CabalSpecV3_8

showCabalSpecVersion :: CabalSpecVersion -> String

Show cabal spec version, but not the way in the .cabal files

Since: Cabal-syntax-3.0.0.0

cabalSpecLatest :: CabalSpecVersion

cabalSpecFromVersionDigits :: [Int] -> Maybe CabalSpecVersion

Parse CabalSpecVersion from version digits.

It may fail if for recent versions the version is not exact.

cabalSpecToVersionDigits :: CabalSpecVersion -> [Int]

Since: Cabal-syntax-3.4.0.0

cabalSpecMinimumLibraryVersion :: CabalSpecVersion -> [Int]

What is the minimum Cabal library version which knows how handle this spec version.

Note: this is a point where we could decouple cabal-spec and Cabal versions, if we ever want that.

>>> cabalSpecMinimumLibraryVersion CabalSpecV3_0
[2,5]

>>> cabalSpecMinimumLibraryVersion CabalSpecV2_4
[2,3]

Since: Cabal-syntax-3.4.0.0

specHasCommonStanzas :: CabalSpecVersion -> HasCommonStanzas

specHasElif :: CabalSpecVersion -> HasElif

data HasElif

Constructors

HasElif

NoElif

data HasCommonStanzas

Constructors

HasCommonStanzas

NoCommonStanzas

data HasGlobstar

Constructors

HasGlobstar

NoGlobstar