Distribution.Compat.CharParsing

Description

Parsers for character streams

Originally in parsers package.

Combinators

oneOf :: CharParsing m => [Char] -> m Char

oneOf cs succeeds if the current character is in the supplied list of characters cs. Returns the parsed character. See also satisfy.

  vowel  = oneOf "aeiou"

noneOf :: CharParsing m => [Char] -> m Char

As the dual of oneOf, noneOf cs succeeds if the current character is not in the supplied list of characters cs. Returns the parsed character.

 consonant = noneOf "aeiou"

spaces :: CharParsing m => m ()

Skips zero or more white space characters. See also skipMany.

space :: CharParsing m => m Char

Parses a white space character (any character which satisfies isSpace) Returns the parsed character.

newline :: CharParsing m => m Char

Parses a newline character ('\n'). Returns a newline character.

tab :: CharParsing m => m Char

Parses a tab character ('\t'). Returns a tab character.

upper :: CharParsing m => m Char

Parses an upper case letter. Returns the parsed character.

lower :: CharParsing m => m Char

Parses a lower case character. Returns the parsed character.

alphaNum :: CharParsing m => m Char

Parses a letter or digit. Returns the parsed character.

letter :: CharParsing m => m Char

Parses a letter (an upper case or lower case character). Returns the parsed character.

digit :: CharParsing m => m Char

Parses a digit. Returns the parsed character.

hexDigit :: CharParsing m => m Char

Parses a hexadecimal digit (a digit or a letter between 'a' and 'f' or 'A' and 'F'). Returns the parsed character.

octDigit :: CharParsing m => m Char

Parses an octal digit (a character between '0' and '7'). Returns the parsed character.

satisfyRange :: CharParsing m => Char -> Char -> m Char

Class

class Parsing m => CharParsing m where

Additional functionality needed to parse character streams.

Minimal complete definition

satisfy

Methods

satisfy :: (Char -> Bool) -> m Char

Parse a single character of the input, with UTF-8 decoding

char :: Char -> m Char

char c parses a single character c. Returns the parsed character (i.e. c).

e.g.

semiColon = char ';'

notChar :: Char -> m Char

notChar c parses any single character other than c. Returns the parsed character.

anyChar :: m Char

This parser succeeds for any character. Returns the parsed character.

string :: String -> m String

string s parses a sequence of characters given by s. Returns the parsed string (i.e. s).

 divOrMod    =   string "div"
             <|> string "mod"

text :: Text -> m Text

text t parses a sequence of characters determined by the text t Returns the parsed text fragment (i.e. t).

Using OverloadedStrings:

 divOrMod    =   text "div"
             <|> text "mod"

Cabal additions

integral :: (CharParsing m, Integral a) => m a

signedIntegral :: (CharParsing m, Integral a) => m a

Accepts negative (starting with -) and positive (without sign) integral numbers.

Since: Cabal-syntax-3.4.0.0

munch1 :: CharParsing m => (Char -> Bool) -> m String

Greedily munch characters while predicate holds. Require at least one character.

munch :: CharParsing m => (Char -> Bool) -> m String

Greedily munch characters while predicate holds. Always succeeds.

skipSpaces1 :: CharParsing m => m ()

module Distribution.Compat.Parsing