Distribution.Compat.Graph

Description

A data type representing directed graphs, backed by Data.Graph. It is strict in the node type.

This is an alternative interface to Data.Graph. In this interface, nodes (identified by the IsNode type class) are associated with a key and record the keys of their neighbors. This interface is more convenient than Graph, which requires vertices to be explicitly handled by integer indexes.

The current implementation has somewhat peculiar performance characteristics. The asymptotics of all map-like operations mirror their counterparts in Data.Map. However, to perform a graph operation, we first must build the Data.Graph representation, an operation that takes O(V + E log V). However, this operation can be amortized across all queries on that particular graph.

Some nodes may be broken, i.e., refer to neighbors which are not stored in the graph. In our graph algorithms, we transparently ignore such edges; however, you can easily query for the broken vertices of a graph using broken (and should, e.g., to ensure that a closure of a graph is well-formed.) It's possible to take a closed subset of a broken graph and get a well-formed graph.

Graph type

data Graph a

A graph of nodes a. The nodes are expected to have instance of class IsNode.

class Ord (Key a) => IsNode a where

The IsNode class is used for datatypes which represent directed graph nodes. A node of type a is associated with some unique key of type Key a; given a node we can determine its key (nodeKey) and the keys of its neighbors (nodeNeighbors).

Associated Types

type Key a

Methods

nodeKey :: a -> Key a

nodeNeighbors :: a -> [Key a]

Query

null :: Graph a -> Bool

O(1). Is the graph empty?

size :: Graph a -> Int

O(1). The number of nodes in the graph.

member :: IsNode a => Key a -> Graph a -> Bool

O(log V). Check if the key is in the graph.

lookup :: IsNode a => Key a -> Graph a -> Maybe a

O(log V). Lookup the node at a key in the graph.

Construction

empty :: IsNode a => Graph a

O(1). The empty graph.

insert :: IsNode a => a -> Graph a -> Graph a

O(log V). Insert a node into a graph.

deleteKey :: IsNode a => Key a -> Graph a -> Graph a

O(log V). Delete the node at a key from the graph.

deleteLookup :: IsNode a => Key a -> Graph a -> (Maybe a, Graph a)

O(log V). Lookup and delete. This function returns the deleted value if it existed.

Combine

unionLeft :: IsNode a => Graph a -> Graph a -> Graph a

O(V + V'). Left-biased union, preferring entries from the first map when conflicts occur.

unionRight :: IsNode a => Graph a -> Graph a -> Graph a

O(V + V'). Right-biased union, preferring entries from the second map when conflicts occur. nodeKey x = nodeKey (f x).

Graph algorithms

stronglyConnComp :: Graph a -> [SCC a]

Ω(V + E). Compute the strongly connected components of a graph. Requires amortized construction of graph.

data SCC vertex

Strongly connected component.

Constructors

AcyclicSCC vertex
A single vertex that is not in any cycle.

NECyclicSCC !(NonEmpty vertex)
A maximal set of mutually reachable vertices.

Since: containers-0.7.0

Bundled Patterns

pattern CyclicSCC :: [vertex] -> SCC vertex
Partial pattern synonym for backward compatibility with containers < 0.7.

cycles :: Graph a -> [[a]]

Ω(V + E). Compute the cycles of a graph. Requires amortized construction of graph.

broken :: Graph a -> [(a, [Key a])]

O(1). Return a list of nodes paired with their broken neighbors (i.e., neighbor keys which are not in the graph). Requires amortized construction of graph.

neighbors :: Graph a -> Key a -> Maybe [a]

Lookup the immediate neighbors from a key in the graph. Requires amortized construction of graph.

revNeighbors :: Graph a -> Key a -> Maybe [a]

Lookup the immediate reverse neighbors from a key in the graph. Requires amortized construction of graph.

closure :: Graph a -> [Key a] -> Maybe [a]

Compute the subgraph which is the closure of some set of keys. Returns Nothing if one (or more) keys are not present in the graph. Requires amortized construction of graph.

revClosure :: Graph a -> [Key a] -> Maybe [a]

Compute the reverse closure of a graph from some set of keys. Returns Nothing if one (or more) keys are not present in the graph. Requires amortized construction of graph.

topSort :: Graph a -> [a]

Topologically sort the nodes of a graph. Requires amortized construction of graph.

revTopSort :: Graph a -> [a]

Reverse topologically sort the nodes of a graph. Requires amortized construction of graph.

Conversions

Maps

toMap :: Graph a -> Map (Key a) a

O(1). Convert a graph into a map from keys to nodes. The resulting map m is guaranteed to have the property that all ((k,n) -> k == nodeKey n) (toList m).

Lists

fromDistinctList :: (IsNode a, Show (Key a)) => [a] -> Graph a

O(V log V). Convert a list of nodes (with distinct keys) into a graph.

toList :: Graph a -> [a]

O(V). Convert a graph into a list of nodes.

keys :: Graph a -> [Key a]

O(V). Convert a graph into a list of keys.

Sets

keysSet :: Graph a -> Set (Key a)

O(V). Convert a graph into a set of keys.

Graphs

toGraph :: Graph a -> (Graph, Vertex -> a, Key a -> Maybe Vertex)

O(1). Convert a graph into a Graph. Requires amortized construction of graph.

Node type

data Node k a

A simple, trivial data type which admits an IsNode instance.

Constructors

N a k [k]

nodeValue :: Node k a -> a

Get the value from a Node.