Distribution.Compat.Semigroup

Description

Compatibility layer for Data.Semigroup

Documentation

class Semigroup a where

The class of semigroups (types with an associative binary operation).

Instances should satisfy the following:

Associativity

x <> (y <> z) = (x <> y) <> z

You can alternatively define sconcat instead of (<>), in which case the laws are:

Unit

sconcat (pure x) = x

Multiplication

sconcat (join xss) = sconcat (fmap sconcat xss)

Since: base-4.9.0.0

Minimal complete definition

(<>) | sconcat

Methods

(<>) :: a -> a -> a infixr 6

An associative operation.

Examples

Expand

>>> [1,2,3] <> [4,5,6]
[1,2,3,4,5,6]

>>> Just [1, 2, 3] <> Just [4, 5, 6]
Just [1,2,3,4,5,6]

>>> putStr "Hello, " <> putStrLn "World!"
Hello, World!

class Semigroup a => Monoid a where

The class of monoids (types with an associative binary operation that has an identity). Instances should satisfy the following:

Right identity

x <> mempty = x

Left identity

mempty <> x = x

Associativity

x <> (y <> z) = (x <> y) <> z (Semigroup law)

Concatenation

mconcat = foldr (<>) mempty

You can alternatively define mconcat instead of mempty, in which case the laws are:

Unit

mconcat (pure x) = x

Multiplication

mconcat (join xss) = mconcat (fmap mconcat xss)

Subclass

mconcat (toList xs) = sconcat xs

The method names refer to the monoid of lists under concatenation, but there are many other instances.

Some types can be viewed as a monoid in more than one way, e.g. both addition and multiplication on numbers. In such cases we often define newtypes and make those instances of Monoid, e.g. Sum and Product.

NOTE: Semigroup is a superclass of Monoid since base-4.11.0.0.

Minimal complete definition

mempty | mconcat

Methods

mempty :: a

Identity of mappend

Examples

Expand

>>> "Hello world" <> mempty
"Hello world"

>>> mempty <> [1, 2, 3]
[1,2,3]

mappend :: a -> a -> a

An associative operation

NOTE: This method is redundant and has the default implementation mappend = (<>) since base-4.11.0.0. Should it be implemented manually, since mappend is a synonym for (<>), it is expected that the two functions are defined the same way. In a future GHC release mappend will be removed from Monoid.

mconcat :: [a] -> a

Fold a list using the monoid.

For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.

>>> mconcat ["Hello", " ", "Haskell", "!"]
"Hello Haskell!"

newtype All

Boolean monoid under conjunction (&&).

All x <> All y = All (x && y)

Examples

Expand

>>> All True <> mempty <> All False)
All {getAll = False}

>>> mconcat (map (\x -> All (even x)) [2,4,6,7,8])
All {getAll = False}

>>> All True <> mempty
All {getAll = True}

Constructors

All

Fields

getAll :: Bool

newtype Any

Boolean monoid under disjunction (||).

Any x <> Any y = Any (x || y)

Examples

Expand

>>> Any True <> mempty <> Any False
Any {getAny = True}

>>> mconcat (map (\x -> Any (even x)) [2,4,6,7,8])
Any {getAny = True}

>>> Any False <> mempty
Any {getAny = False}

Constructors

Any

Fields

getAny :: Bool

newtype First' a

A copy of First.

Constructors

First'

Fields

getFirst' :: a

newtype Last' a

A copy of Last.

Constructors

Last'

Fields

getLast' :: a

newtype Option' a

A wrapper around Maybe, providing the Semigroup and Monoid instances implemented for Maybe since base-4.11.

Constructors

Option'

Fields

getOption' :: Maybe a

gmappend :: (Generic a, GSemigroup (Rep a)) => a -> a -> a

Generically generate a Semigroup (<>) operation for any type implementing Generic. This operation will append two values by point-wise appending their component fields. It is only defined for product types.

gmappend a (gmappend b c) = gmappend (gmappend a b) c

gmempty :: (Generic a, GMonoid (Rep a)) => a

Generically generate a Monoid mempty for any product-like type implementing Generic.

It is only defined for product types.

gmappend gmempty a = a = gmappend a gmempty