Distribution.Compat.Typeable

Documentation

class Typeable (a :: k)

The class Typeable allows a concrete representation of a type to be calculated.

Minimal complete definition

typeRep#

type TypeRep = SomeTypeRep

A quantified type representation.

typeRep :: forall {k} proxy (a :: k). Typeable a => proxy a -> TypeRep

Takes a value of type a and returns a concrete representation of that type.

Since: base-4.7.0.0