Distribution.FieldGrammar

Description

This module provides a way to specify a grammar of .cabal -like files.

Field grammar type

class (c SpecVersion, c TestedWith, c SpecLicense, c Token, c Token', c FilePathNT) => FieldGrammar c g | g -> c where

FieldGrammar is parametrised by

s which is a structure we are parsing. We need this to provide prettyprinter functionality

a type of the field.

Note: We'd like to have forall s. Applicative (f s) context.

Minimal complete definition

blurFieldGrammar, uniqueFieldAla, booleanFieldDef, optionalFieldAla, optionalFieldDefAla, freeTextField, freeTextFieldDef, freeTextFieldDefST, monoidalFieldAla, prefixedFields, knownField, hiddenField, deprecatedSince, removedIn, availableSince

Methods

blurFieldGrammar :: ALens' a b -> g b d -> g a d

Unfocus, zoom out, blur FieldGrammar.

uniqueFieldAla

Arguments

:: (c b, Newtype a b)

=> FieldName
field name

-> (a -> b)
Newtype pack

-> ALens' s a
lens into the field

-> g s a

Field which should be defined, exactly once.

booleanFieldDef

Arguments

:: FieldName
field name

-> ALens' s Bool
lens into the field

-> Bool
default

-> g s Bool

Boolean field with a default value.

optionalFieldAla

Arguments

:: (c b, Newtype a b)

=> FieldName
field name

-> (a -> b)
pack

-> ALens' s (Maybe a)
lens into the field

-> g s (Maybe a)

Optional field.

optionalFieldDefAla

Arguments

:: (c b, Newtype a b, Eq a)

=> FieldName
field name

-> (a -> b)
Newtype pack

-> ALens' s a
Lens' s a: lens into the field

-> a
default value

-> g s a

Optional field with default value.

freeTextField

Arguments

:: FieldName

-> ALens' s (Maybe String)
lens into the field

-> g s (Maybe String)

freeTextFieldDef

Arguments

:: FieldName

-> ALens' s String
lens into the field

-> g s String

freeTextFieldDefST

Arguments

:: FieldName

-> ALens' s ShortText
lens into the field

-> g s ShortText

Since: Cabal-syntax-3.2.0.0

monoidalFieldAla

Arguments

:: (c b, Monoid a, Newtype a b)

=> FieldName
field name

-> (a -> b)
pack

-> ALens' s a
lens into the field

-> g s a

Monoidal field.

Values are combined with mappend.

Note: optionalFieldAla is a monoidalField with Last monoid.

prefixedFields

Arguments

:: FieldName
field name prefix

-> ALens' s [(String, String)]
lens into the field

-> g s [(String, String)]

Parser matching all fields with a name starting with a prefix.

knownField :: FieldName -> g s ()

Known field, which we don't parse, nor pretty print.

hiddenField :: g s a -> g s a

Field which is parsed but not pretty printed.

deprecatedSince

Arguments

:: CabalSpecVersion
version

-> String
deprecation message

-> g s a

-> g s a

Deprecated since

removedIn

Arguments

:: CabalSpecVersion
version

-> String
removal message

-> g s a

-> g s a

Removed in. If we encounter removed field, parsing fails.

availableSince

Arguments

:: CabalSpecVersion
spec version

-> a
default value

-> g s a

-> g s a

Annotate field with since spec-version.

availableSinceWarn

Arguments

:: CabalSpecVersion
spec version

-> g s a

-> g s a

Annotate field with since spec-version. This is used to recognise, but warn about the field. It is used to process other-extensions field.

Default implementation is to not warn.

Since: Cabal-syntax-3.4.0.0

uniqueField

Arguments

:: (FieldGrammar c g, c (Identity a))

=> FieldName
field name

-> ALens' s a
lens into the field

-> g s a

Field which can be defined at most once.

optionalField

Arguments

:: (FieldGrammar c g, c (Identity a))

=> FieldName
field name

-> ALens' s (Maybe a)
lens into the field

-> g s (Maybe a)

Field which can be defined at most once.

optionalFieldDef

Arguments

:: (FieldGrammar c g, Functor (g s), c (Identity a), Eq a)

=> FieldName
field name

-> ALens' s a
Lens' s a: lens into the field

-> a
default value

-> g s a

Optional field with default value.

monoidalField

Arguments

:: (FieldGrammar c g, c (Identity a), Monoid a)

=> FieldName
field name

-> ALens' s a
lens into the field

-> g s a

Field which can be define multiple times, and the results are mappended.

Concrete grammar implementations

data ParsecFieldGrammar s a

type ParsecFieldGrammar' a = ParsecFieldGrammar a a

parseFieldGrammar :: CabalSpecVersion -> Fields Position -> ParsecFieldGrammar s a -> ParseResult a

fieldGrammarKnownFieldList :: ParsecFieldGrammar s a -> [FieldName]

data PrettyFieldGrammar s a

type PrettyFieldGrammar' a = PrettyFieldGrammar a a

prettyFieldGrammar :: CabalSpecVersion -> PrettyFieldGrammar s a -> s -> [PrettyField ()]

We can use PrettyFieldGrammar to pp print the s.

Note: there is not trailing ($+$ text "").

Auxiliary

(^^^) :: a -> (a -> b) -> b infixl 5

Reverse function application which binds tighter than <$> and <*>. Useful for refining grammar specification.

<*> monoidalFieldAla "extensions"           (alaList' FSep MQuoted)       oldExtensions
    ^^^ deprecatedSince [1,12] "Please use 'default-extensions' or 'other-extensions' fields."

data Section ann

The Section constructor of Field.

Constructors

MkSection !(Name ann) [SectionArg ann] [Field ann]

type Fields ann = Map FieldName [NamelessField ann]

partitionFields :: [Field ann] -> (Fields ann, [[Section ann]])

Partition field list into field map and groups of sections.

takeFields :: [Field ann] -> (Fields ann, [Field ann])

Take all fields from the front.

runFieldParser :: Position -> ParsecParser a -> CabalSpecVersion -> [FieldLine Position] -> ParseResult a

runFieldParser' :: [Position] -> ParsecParser a -> CabalSpecVersion -> FieldLineStream -> ParseResult a

defaultFreeTextFieldDefST

Arguments

:: (Functor (g s), FieldGrammar c g)

=> FieldName

-> ALens' s ShortText
lens into the field

-> g s ShortText

Default implementation for freeTextFieldDefST.

Newtypes

module Distribution.FieldGrammar.Newtypes