Distribution.FieldGrammar.Newtypes

Description

This module provides newtype wrappers to be used with Distribution.FieldGrammar.

List

alaList :: sep -> [a] -> List sep (Identity a) a

alaList and alaList' are simply List, with additional phantom arguments to constrain the resulting type

>>> :t alaList VCat
alaList VCat :: [a] -> List VCat (Identity a) a

>>> :t alaList' FSep Token
alaList' FSep Token :: [String] -> List FSep Token String

alaList' :: sep -> (a -> b) -> [a] -> List sep b a

More general version of alaList.

Modifiers

data CommaVCat

Vertical list with commas. Displayed with vcat

Constructors

CommaVCat

data CommaFSep

Paragraph fill list with commas. Displayed with fsep

Constructors

CommaFSep

data VCat

Vertical list with optional commas. Displayed with vcat.

Constructors

VCat

data FSep

Paragraph fill list with optional commas. Displayed with fsep.

Constructors

FSep

data NoCommaFSep

Paragraph fill list without commas. Displayed with fsep.

Constructors

NoCommaFSep

class Sep sep where

Methods

prettySep :: Proxy sep -> [Doc] -> Doc

parseSep :: CabalParsing m => Proxy sep -> m a -> m [a]

parseSepNE :: CabalParsing m => Proxy sep -> m a -> m (NonEmpty a)

Type

data List sep b a

List separated with optional commas. Displayed with sep, arguments of type a are parsed and pretty-printed as b.

Set

alaSet :: sep -> Set a -> Set' sep (Identity a) a

alaSet and alaSet' are simply Set' constructor, with additional phantom arguments to constrain the resulting type

>>> :t alaSet VCat
alaSet VCat :: Set a -> Set' VCat (Identity a) a

>>> :t alaSet' FSep Token
alaSet' FSep Token :: Set String -> Set' FSep Token String

>>> unpack' (alaSet' FSep Token) <$> eitherParsec "foo bar foo"
Right (fromList ["bar","foo"])

Since: Cabal-syntax-3.2.0.0

alaSet' :: sep -> (a -> b) -> Set a -> Set' sep b a

More general version of alaSet.

Since: Cabal-syntax-3.2.0.0

data Set' sep b a

Like List, but for Set.

Since: Cabal-syntax-3.2.0.0

NonEmpty

alaNonEmpty :: sep -> NonEmpty a -> NonEmpty' sep (Identity a) a

alaNonEmpty and alaNonEmpty' are simply NonEmpty' constructor, with additional phantom arguments to constrain the resulting type

>>> :t alaNonEmpty VCat
alaNonEmpty VCat :: NonEmpty a -> NonEmpty' VCat (Identity a) a

>>> unpack' (alaNonEmpty' FSep Token) <$> eitherParsec "foo bar foo"
Right ("foo" :| ["bar","foo"])

Since: Cabal-syntax-3.2.0.0

alaNonEmpty' :: sep -> (a -> b) -> NonEmpty a -> NonEmpty' sep b a

More general version of alaNonEmpty.

Since: Cabal-syntax-3.2.0.0

data NonEmpty' sep b a

Like List, but for NonEmpty.

Since: Cabal-syntax-3.2.0.0

Version & License

newtype SpecVersion

Version range or just version, i.e. cabal-version field.

There are few things to consider:

Starting with 2.2 the cabal-version field should be the first field in the file and only exact version is accepted. Therefore if we get e.g. >= 2.2, we fail. See https://github.com/haskell/cabal/issues/4899

We have this newtype, as writing Parsec and Pretty instances for CabalSpecVersion would cause cycle in modules: Version -> CabalSpecVersion -> Parsec -> ...

Constructors

SpecVersion

Fields

getSpecVersion :: CabalSpecVersion

newtype TestedWith

Version range or just version

Constructors

TestedWith

Fields

getTestedWith :: (CompilerFlavor, VersionRange)

newtype SpecLicense

SPDX License expression or legacy license

Constructors

SpecLicense

Fields

getSpecLicense :: Either License License

Identifiers

newtype Token

Haskell string or [^ ,]+

Constructors

Token

Fields

getToken :: String

newtype Token'

Haskell string or [^ ]+

Constructors

Token'

Fields

getToken' :: String

newtype MQuoted a

Either "quoted" or un-quoted.

Constructors

MQuoted

Fields

getMQuoted :: a

newtype FilePathNT

Filepath are parsed as Token.

Constructors

FilePathNT

Fields

getFilePathNT :: String