Distribution.Fields

Description

Utilities to work with .cabal like file structure.

Types

data Field ann

A Cabal-like file consists of a series of fields (foo: bar) and sections (library ...).

Constructors

Field !(Name ann) [FieldLine ann]

Section !(Name ann) [SectionArg ann] [Field ann]

data Name ann

A field name.

Invariant: ByteString is lower-case ASCII.

Constructors

Name !ann !FieldName

data FieldLine ann

A line of text representing the value of a field from a Cabal file. A field may contain multiple lines.

Invariant: ByteString has no newlines.

Constructors

FieldLine !ann !ByteString

data SectionArg ann

Section arguments, e.g. name of the library

Constructors

SecArgName !ann !ByteString
identifier, or something which looks like number. Also many dot numbers, i.e. "7.6.3"

SecArgStr !ann !ByteString
quoted string

SecArgOther !ann !ByteString
everything else, mm. operators (e.g. in if-section conditionals)

type FieldName = ByteString

Grammar and parsing

readFields :: ByteString -> Either ParseError [Field Position]

Parse cabal style ByteString into list of Fields, i.e. the cabal AST.

readFields' :: ByteString -> Either ParseError ([Field Position], [LexWarning])

Like readFields but also return lexer warnings

ParseResult

data ParseResult a

A monad with failure and accumulating errors and warnings.

runParseResult :: ParseResult a -> ([PWarning], Either (Maybe Version, NonEmpty PError) a)

Destruct a ParseResult into the emitted warnings and either a successful value or list of errors and possibly recovered a spec-version declaration.

parseWarning :: Position -> PWarnType -> String -> ParseResult ()

Add a warning. This doesn't fail the parsing process.

parseWarnings :: [PWarning] -> ParseResult ()

Add multiple warnings at once.

parseFailure :: Position -> String -> ParseResult ()

Add an error, but not fail the parser yet.

For fatal failure use parseFatalFailure

parseFatalFailure :: Position -> String -> ParseResult a

Add an fatal error.

Warnings

data PWarnType

Type of parser warning. We do classify warnings.

Different application may decide not to show some, or have fatal behaviour on others

Constructors

PWTOther
Unclassified warning

PWTUTF
Invalid UTF encoding

PWTBoolCase
true or false, not True or False

PWTVersionTag
there are version with tags

PWTNewSyntax
New syntax used, but no cabal-version: >= 1.2 specified

PWTOldSyntax
Old syntax used, and cabal-version >= 1.2 specified

PWTDeprecatedField

PWTInvalidSubsection

PWTUnknownField

PWTUnknownSection

PWTTrailingFields

PWTExtraMainIs
extra main-is field

PWTExtraTestModule
extra test-module field

PWTExtraBenchmarkModule
extra benchmark-module field

PWTLexNBSP

PWTLexBOM

PWTLexTab

PWTQuirkyCabalFile
legacy cabal file that we know how to patch

PWTDoubleDash
Double dash token, most likely it's a mistake - it's not a comment

PWTMultipleSingularField
e.g. name or version should be specified only once.

PWTBuildTypeDefault
Workaround for derive-package having build-type: Default. See https://github.com/haskell/cabal/issues/5020.

PWTVersionOperator
Version operators used (without cabal-version: 1.8)

PWTVersionWildcard
Version wildcard used (without cabal-version: 1.6)

PWTSpecVersion
Warnings about cabal-version format.

PWTEmptyFilePath
Empty filepath, i.e. literally ""

PWTExperimental
Experimental feature

data PWarning

Parser warning.

Constructors

PWarning !PWarnType !Position String

showPWarning :: FilePath -> PWarning -> String

Errors

data PError

Parser error.

Constructors

PError Position String

showPError :: FilePath -> PError -> String

Pretty printing

data CommentPosition

This type is used to discern when a comment block should go before or after a cabal-like file field, otherwise it would be hardcoded to a single position. It is often used in conjunction with PrettyField.

Constructors

CommentBefore [String]

CommentAfter [String]

NoComment

data PrettyField ann

Constructors

PrettyField ann FieldName Doc

PrettySection ann FieldName [Doc] [PrettyField ann]

PrettyEmpty

showFields :: (ann -> CommentPosition) -> [PrettyField ann] -> String

Prettyprint a list of fields.

Note: the first argument should return Strings without newlines and properly prefixes (with --) to count as comments. This unsafety is left in place so one could generate empty lines between comment lines.

Transformation from Field

genericFromParsecFields

Arguments

:: Applicative f

=> (FieldName -> [FieldLine ann] -> f Doc)
transform field contents

-> (FieldName -> [SectionArg ann] -> f [Doc])
transform section arguments

-> [Field ann]

-> f [PrettyField ann]

fromParsecFields :: [Field ann] -> [PrettyField ann]

Simple variant of genericFromParsecField