Distribution.License

Description

Package descriptions contain fields for specifying the name of a software license and the name of the file containing the text of that license. While package authors may choose any license they like, Cabal provides an enumeration of a small set of common free and open source software licenses. This is done so that Hackage can recognise licenses, so that tools can detect licensing conflicts, and to deter license proliferation.

It is recommended that all package authors use the license-file or license-files fields in their package descriptions. Further information about these fields can be found in the Cabal users guide.

Additional resources

The following websites provide information about free and open source software licenses:

The Open Source Initiative (OSI)

The Free Software Foundation (FSF)

Disclaimer

The descriptions of software licenses provided by this documentation are intended for informational purposes only and in no way constitute legal advice. Please read the text of the licenses and consult a lawyer for any advice regarding software licensing.

Documentation

data License

Indicates the license under which a package's source code is released. Versions of the licenses not listed here will be rejected by Hackage and cause cabal check to issue a warning.

Constructors

GPL (Maybe Version)
GNU General Public License, version 2 or version 3.

AGPL (Maybe Version)
GNU Affero General Public License, version 3.

LGPL (Maybe Version)
GNU Lesser General Public License, version 2.1 or version 3.

BSD2
2-clause BSD license.

BSD3
3-clause BSD license.

BSD4
4-clause BSD license. This license has not been approved by the OSI and is incompatible with the GNU GPL. It is provided for historical reasons and should be avoided.

MIT
MIT license.

ISC
ISC license

MPL Version
Mozilla Public License, version 2.0.

Apache (Maybe Version)
Apache License, version 2.0.

PublicDomain
The author of a package disclaims any copyright to its source code and dedicates it to the public domain. This is not a software license. Please note that it is not possible to dedicate works to the public domain in every jurisdiction, nor is a work that is in the public domain in one jurisdiction necessarily in the public domain elsewhere.

AllRightsReserved
Explicitly 'All Rights Reserved', eg for proprietary software. The package may not be legally modified or redistributed by anyone but the rightsholder.

UnspecifiedLicense
No license specified which legally defaults to 'All Rights Reserved'. The package may not be legally modified or redistributed by anyone but the rightsholder.

OtherLicense
Any other software license.

UnknownLicense String
Indicates an erroneous license name.

knownLicenses :: [License]

The list of all currently recognised licenses.

licenseToSPDX :: License -> License

Convert old License to SPDX License. Non-SPDX licenses are converted to LicenseRef.

Since: Cabal-syntax-2.2.0.0

licenseFromSPDX :: License -> License

Convert License to License,

This is lossy conversion. We try our best.

>>> licenseFromSPDX . licenseToSPDX $ BSD3
BSD3

>>> licenseFromSPDX . licenseToSPDX $ GPL (Just (mkVersion [3]))
GPL (Just (mkVersion [3]))

>>> licenseFromSPDX . licenseToSPDX $ PublicDomain
UnknownLicense "LicenseRefPublicDomain"

>>> licenseFromSPDX $ SPDX.License $ SPDX.simpleLicenseExpression SPDX.EUPL_1_1
UnknownLicense "EUPL-1.1"

>>> licenseFromSPDX . licenseToSPDX $ AllRightsReserved
AllRightsReserved

>>> licenseFromSPDX <$> simpleParsec "BSD-3-Clause OR GPL-3.0-only"
Just (UnknownLicense "BSD3ClauseORGPL30only")

Since: Cabal-syntax-2.2.0.0