Distribution.Package

Description

Defines a package identifier along with a parser and pretty printer for it. PackageIdentifiers consist of a name and an exact version. It also defines a Dependency data type. A dependency is a package name and a version range, like "foo >= 1.2 && < 2".

Documentation

module Distribution.Types.AbiHash

module Distribution.Types.ComponentId

module Distribution.Types.PackageId

module Distribution.Types.UnitId

module Distribution.Types.Module

module Distribution.Types.PackageName

module Distribution.Types.PkgconfigName

module Distribution.Types.Dependency

class Package pkg where

Class of things that have a PackageIdentifier

Types in this class are all notions of a package. This allows us to have different types for the different phases that packages go though, from simple name/id, package description, configured or installed packages.

Not all kinds of packages can be uniquely identified by a PackageIdentifier. In particular, installed packages cannot, there may be many installed instances of the same source package.

Methods

packageId :: pkg -> PackageIdentifier

packageName :: Package pkg => pkg -> PackageName

packageVersion :: Package pkg => pkg -> Version

class HasMungedPackageId pkg where

Methods

mungedId :: pkg -> MungedPackageId

mungedName' :: HasMungedPackageId pkg => pkg -> MungedPackageName

mungedVersion' :: HasMungedPackageId munged => munged -> Version

class Package pkg => HasUnitId pkg where

Packages that have an installed unit ID

Methods

installedUnitId :: pkg -> UnitId

class HasUnitId pkg => PackageInstalled pkg where

Class of installed packages.

The primary data type which is an instance of this package is InstalledPackageInfo, but when we are doing install plans in Cabal install we may have other, installed package-like things which contain more metadata. Installed packages have exact dependencies installedDepends.

Methods

installedDepends :: pkg -> [UnitId]