Distribution.PackageDescription.Parsec

Description

This defined parsers and partial pretty printers for the .cabal format.

Package descriptions

parseGenericPackageDescription :: ByteString -> ParseResult GenericPackageDescription

Parses the given file into a GenericPackageDescription.

In Cabal 1.2 the syntax for package descriptions was changed to a format with sections and possibly indented property descriptions.

parseGenericPackageDescriptionMaybe :: ByteString -> Maybe GenericPackageDescription

Maybe variant of parseGenericPackageDescription

Parsing

data ParseResult a

A monad with failure and accumulating errors and warnings.

runParseResult :: ParseResult a -> ([PWarning], Either (Maybe Version, NonEmpty PError) a)

Destruct a ParseResult into the emitted warnings and either a successful value or list of errors and possibly recovered a spec-version declaration.

New-style spec-version

scanSpecVersion :: ByteString -> Maybe Version

Quickly scan new-style spec-version

A new-style spec-version declaration begins the .cabal file and follow the following case-insensitive grammar (expressed in RFC5234 ABNF):

newstyle-spec-version-decl = "cabal-version" *WS ":" *WS newstyle-pec-version *WS

spec-version               = NUM "." NUM [ "." NUM ]

NUM    = DIGIT0 / DIGITP 1*DIGIT0
DIGIT0 = %x30-39
DIGITP = %x31-39
WS = %20

Supplementary build information

parseHookedBuildInfo :: ByteString -> ParseResult HookedBuildInfo