Distribution.Parsec

Documentation

class Parsec a where

Class for parsing with parsec. Mainly used for .cabal file fields.

For parsing .cabal like file structure, see Distribution.Fields.

Methods

parsec :: CabalParsing m => m a

newtype ParsecParser a

Constructors

PP

Fields

unPP :: CabalSpecVersion -> Parsec FieldLineStream [PWarning] a

runParsecParser :: ParsecParser a -> FilePath -> FieldLineStream -> Either ParseError a

Run ParsecParser with cabalSpecLatest.

runParsecParser' :: CabalSpecVersion -> ParsecParser a -> FilePath -> FieldLineStream -> Either ParseError a

Like runParsecParser but lets specify CabalSpecVersion used.

Since: Cabal-syntax-3.0.0.0

simpleParsec :: Parsec a => String -> Maybe a

Parse a String with lexemeParsec.

simpleParsecBS :: Parsec a => ByteString -> Maybe a

Like simpleParsec but for ByteString

simpleParsec' :: Parsec a => CabalSpecVersion -> String -> Maybe a

Parse a String with lexemeParsec using specific CabalSpecVersion.

Since: Cabal-syntax-3.4.0.0

simpleParsecW' :: Parsec a => CabalSpecVersion -> String -> Maybe a

Parse a String with lexemeParsec using specific CabalSpecVersion. Fail if there are any warnings.

Since: Cabal-syntax-3.4.0.0

lexemeParsec :: (CabalParsing m, Parsec a) => m a

parsec could consume trailing spaces, this function will consume.

eitherParsec :: Parsec a => String -> Either String a

Parse a String with lexemeParsec.

explicitEitherParsec :: ParsecParser a -> String -> Either String a

Parse a String with given ParsecParser. Trailing whitespace is accepted.

explicitEitherParsec' :: CabalSpecVersion -> ParsecParser a -> String -> Either String a

Parse a String with given ParsecParser and CabalSpecVersion. Trailing whitespace is accepted. See explicitEitherParsec.

Since: Cabal-syntax-3.4.0.0

CabalParsing and diagnostics

class (CharParsing m, MonadPlus m, MonadFail m) => CabalParsing m where

Parsing class which

can report Cabal parser warnings.

knows cabal-version we work with

Minimal complete definition

parsecWarning, askCabalSpecVersion

Methods

parsecWarning :: PWarnType -> String -> m ()

parsecHaskellString :: m String

askCabalSpecVersion :: m CabalSpecVersion

Warnings

data PWarnType

Type of parser warning. We do classify warnings.

Different application may decide not to show some, or have fatal behaviour on others

Constructors

PWTOther
Unclassified warning

PWTUTF
Invalid UTF encoding

PWTBoolCase
true or false, not True or False

PWTVersionTag
there are version with tags

PWTNewSyntax
New syntax used, but no cabal-version: >= 1.2 specified

PWTOldSyntax
Old syntax used, and cabal-version >= 1.2 specified

PWTDeprecatedField

PWTInvalidSubsection

PWTUnknownField

PWTUnknownSection

PWTTrailingFields

PWTExtraMainIs
extra main-is field

PWTExtraTestModule
extra test-module field

PWTExtraBenchmarkModule
extra benchmark-module field

PWTLexNBSP

PWTLexBOM

PWTLexTab

PWTQuirkyCabalFile
legacy cabal file that we know how to patch

PWTDoubleDash
Double dash token, most likely it's a mistake - it's not a comment

PWTMultipleSingularField
e.g. name or version should be specified only once.

PWTBuildTypeDefault
Workaround for derive-package having build-type: Default. See https://github.com/haskell/cabal/issues/5020.

PWTVersionOperator
Version operators used (without cabal-version: 1.8)

PWTVersionWildcard
Version wildcard used (without cabal-version: 1.6)

PWTSpecVersion
Warnings about cabal-version format.

PWTEmptyFilePath
Empty filepath, i.e. literally ""

PWTExperimental
Experimental feature

data PWarning

Parser warning.

Constructors

PWarning !PWarnType !Position String

showPWarning :: FilePath -> PWarning -> String

Errors

data PError

Parser error.

Constructors

PError Position String

showPError :: FilePath -> PError -> String

Position

data Position

1-indexed row and column positions in a file.

Constructors

Position !Int !Int

incPos :: Int -> Position -> Position

Shift position by n columns to the right.

retPos :: Position -> Position

Shift position to beginning of next row.

showPos :: Position -> String

zeroPos :: Position

Utilities

parsecToken :: CabalParsing m => m String

[^ ,]

parsecToken' :: CabalParsing m => m String

[^ ]

parsecFilePath :: CabalParsing m => m FilePath

parsecQuoted :: CabalParsing m => m a -> m a

Content isn't unquoted

parsecMaybeQuoted :: CabalParsing m => m a -> m a

parsecMaybeQuoted p = parsecQuoted p | p.

parsecCommaList :: CabalParsing m => m a -> m [a]

parsecCommaNonEmpty :: CabalParsing m => m a -> m (NonEmpty a)

parsecLeadingCommaList :: CabalParsing m => m a -> m [a]

Like parsecCommaList but accept leading or trailing comma.

p (comma p)*  -- p sepBy comma
(comma p)*    -- leading comma
(p comma)*    -- trailing comma

parsecLeadingCommaNonEmpty :: CabalParsing m => m a -> m (NonEmpty a)

Since: Cabal-syntax-3.4.0.0

parsecOptCommaList :: CabalParsing m => m a -> m [a]

parsecLeadingOptCommaList :: CabalParsing m => m a -> m [a]

Like parsecOptCommaList but

require all or none commas

accept leading or trailing comma.

p (comma p)*  -- p sepBy comma
(comma p)*    -- leading comma
(p comma)*    -- trailing comma
p*            -- no commas: many p

Since: Cabal-syntax-3.0.0.0

parsecStandard :: (CabalParsing m, Parsec ver) => (ver -> String -> a) -> m a

Parse a benchmark/test-suite types.

parsecUnqualComponentName :: forall m. CabalParsing m => m String