Distribution.System

Description

Cabal often needs to do slightly different things on specific platforms. You probably know about the os however using that is very inconvenient because it is a string and different Haskell implementations do not agree on using the same strings for the same platforms! (In particular see the controversy over "windows" vs "mingw32"). So to make it more consistent and easy to use we have an OS enumeration.

Operating System

data OS

These are the known OS names: Linux, Windows, OSX ,FreeBSD, OpenBSD, NetBSD, DragonFly ,Solaris, AIX, HPUX, IRIX ,HaLVM ,Hurd ,IOS, Android, Ghcjs, Wasi

The following aliases can also be used:, * Windows aliases: mingw32, win32, cygwin32 * OSX alias: darwin * Hurd alias: gnu * FreeBSD alias: kfreebsdgnu * Solaris alias: solaris2

Constructors

Linux

Windows

OSX

FreeBSD

OpenBSD

NetBSD

DragonFly

Solaris

AIX

HPUX

IRIX

HaLVM

Hurd

IOS

Android

Ghcjs

Wasi

OtherOS String

buildOS :: OS

Machine Architecture

data Arch

These are the known Arches: I386, X86_64, PPC, PPC64, Sparc, Arm, AArch64, Mips, SH, IA64, S390, S390X, Alpha, Hppa, Rs6000, M68k, Vax, JavaScript and Wasm32.

The following aliases can also be used: * PPC alias: powerpc * PPC64 alias : powerpc64, powerpc64le * Sparc aliases: sparc64, sun4 * Mips aliases: mipsel, mipseb * Arm aliases: armeb, armel * AArch64 aliases: arm64

Constructors

I386

X86_64

PPC

PPC64

Sparc

Arm

AArch64

Mips

SH

IA64

S390

S390X

Alpha

Hppa

Rs6000

M68k

Vax

JavaScript

Wasm32

OtherArch String

buildArch :: Arch

Platform is a pair of arch and OS

data Platform

Constructors

Platform Arch OS

buildPlatform :: Platform

The platform Cabal was compiled on. In most cases, LocalBuildInfo.hostPlatform should be used instead (the platform we're targeting).

platformFromTriple :: String -> Maybe Platform

Internal

knownOSs :: [OS]

knownArches :: [Arch]

Classification

data ClassificationStrictness

How strict to be when classifying strings into the OS and Arch enums.

The reason we have multiple ways to do the classification is because there are two situations where we need to do it.

For parsing OS and arch names in .cabal files we really want everyone to be referring to the same or arch by the same name. Variety is not a virtue in this case. We don't mind about case though.

For the System.Info.os/arch different Haskell implementations use different names for the same or/arch. Also they tend to distinguish versions of an OS/arch which we just don't care about.

The Compat classification allows us to recognise aliases that are already in common use but it allows us to distinguish them from the canonical name which enables us to warn about such deprecated aliases.

Constructors

Permissive

Compat

Strict

classifyOS :: ClassificationStrictness -> String -> OS

classifyArch :: ClassificationStrictness -> String -> Arch