Distribution.Types.BuildType

Documentation

data BuildType

The type of build system used by this package.

Constructors

Simple
calls Distribution.Simple.defaultMain

Configure
calls Distribution.Simple.defaultMainWithHooks defaultUserHooks, which invokes configure to generate additional build information used by later phases.

Make
calls Distribution.Make.defaultMain

Custom
uses user-supplied Setup.hs or Setup.lhs (default)

knownBuildTypes :: [BuildType]