Distribution.Types.Dependency

Documentation

data Dependency

Describes a dependency on a source package (API)

Invariant: package name does not appear as LSubLibName in set of library names.

Note: Dependency is not an instance of Ord, and so it cannot be used in Set or as the key to a Map. For these and similar use cases see DependencyMap.

Constructors

Dependency PackageName VersionRange (NonEmptySet LibraryName)
The set of libraries required from the package. Only the selected libraries will be built. It does not affect the cabal-install solver yet.

mkDependency :: PackageName -> VersionRange -> NonEmptySet LibraryName -> Dependency

Smart constructor of Dependency.

If PackageName is appears as LSubLibName in a set of sublibraries, it is automatically converted to LMainLibName.

Since: Cabal-syntax-3.4.0.0

depPkgName :: Dependency -> PackageName

depVerRange :: Dependency -> VersionRange

depLibraries :: Dependency -> NonEmptySet LibraryName

simplifyDependency :: Dependency -> Dependency

Simplify the VersionRange expression in a Dependency. See simplifyVersionRange.

mainLibSet :: NonEmptySet LibraryName

Library set with main library.

Since: Cabal-syntax-3.4.0.0