Distribution.Types.Flag

Package flag

data PackageFlag

A flag can represent a feature to be included, or a way of linking a target against its dependencies, or in fact whatever you can think of.

Constructors

MkPackageFlag

Fields

flagName :: FlagName

flagDescription :: String

flagDefault :: Bool

flagManual :: Bool

emptyFlag :: FlagName -> PackageFlag

A PackageFlag initialized with default parameters.

Flag name

data FlagName

A FlagName is the name of a user-defined configuration flag

Use mkFlagName and unFlagName to convert from/to a String.

This type is opaque since Cabal-2.0

Since: Cabal-syntax-2.0.0.2

mkFlagName :: String -> FlagName

Construct a FlagName from a String

mkFlagName is the inverse to unFlagName

Note: No validations are performed to ensure that the resulting FlagName is valid

Since: Cabal-syntax-2.0.0.2

unFlagName :: FlagName -> String

Convert FlagName to String

Since: Cabal-syntax-2.0.0.2

Flag assignment

data FlagAssignment

A FlagAssignment is a total or partial mapping of FlagNames to Bool flag values. It represents the flags chosen by the user or discovered during configuration. For example --flags=foo --flags=-bar becomes [("foo", True), ("bar", False)]

TODO: Why we record the multiplicity of the flag?

mkFlagAssignment :: [(FlagName, Bool)] -> FlagAssignment

Construct a FlagAssignment from a list of flag/value pairs.

If duplicate flags occur in the input list, the later entries in the list will take precedence.

Since: Cabal-syntax-2.2.0

unFlagAssignment :: FlagAssignment -> [(FlagName, Bool)]

Deconstruct a FlagAssignment into a list of flag/value pairs.

 null (findDuplicateFlagAssignments fa) ==> (mkFlagAssignment . unFlagAssignment) fa == fa

Since: Cabal-syntax-2.2.0

lookupFlagAssignment :: FlagName -> FlagAssignment -> Maybe Bool

Lookup the value for a flag

Returns Nothing if the flag isn't contained in the FlagAssignment.

Since: Cabal-syntax-2.2.0

insertFlagAssignment :: FlagName -> Bool -> FlagAssignment -> FlagAssignment

Insert or update the boolean value of a flag.

If the flag is already present in the FlagAssignment, the value will be updated and the fact that multiple values have been provided for that flag will be recorded so that a warning can be generated later on.

Since: Cabal-syntax-2.2.0

diffFlagAssignment :: FlagAssignment -> FlagAssignment -> FlagAssignment

Remove all flag-assignments from the first FlagAssignment that are contained in the second FlagAssignment

NB/TODO: This currently only removes flag assignments which also match the value assignment! We should review the code which uses this operation to figure out if this it's not enough to only compare the flagnames without the values.

Since: Cabal-syntax-2.2.0

findDuplicateFlagAssignments :: FlagAssignment -> [FlagName]

Find the FlagNames that have been listed more than once.

Since: Cabal-syntax-2.2.0

nullFlagAssignment :: FlagAssignment -> Bool

Test whether FlagAssignment is empty.

Since: Cabal-syntax-2.2.0

showFlagValue :: (FlagName, Bool) -> String

String representation of a flag-value pair.

dispFlagAssignment :: FlagAssignment -> Doc

Pretty-prints a flag assignment.

showFlagAssignment :: FlagAssignment -> String

Show flag assignment.

Since: Cabal-syntax-3.4.0.0

parsecFlagAssignment :: CabalParsing m => m FlagAssignment

Parses a flag assignment.

parsecFlagAssignmentNonEmpty :: CabalParsing m => m FlagAssignment

Parse a non-empty flag assignment

The flags have to explicitly start with minus or plus.

Since: Cabal-syntax-3.4.0.0

Legacy formats

legacyShowFlagAssignment :: FlagAssignment -> String

We need this as far as we support custom setups older than 2.2.0.0

Since: Cabal-syntax-3.4.0.0

legacyShowFlagAssignment' :: FlagAssignment -> [String]

Since: Cabal-syntax-3.4.0.0

legacyParsecFlagAssignment :: CabalParsing m => m FlagAssignment

We need this as far as we support custom setups older than 2.2.0.0

Since: Cabal-syntax-3.4.0.0