Distribution.Types.ForeignLib.Lens

Documentation

data ForeignLib

A foreign library stanza is like a library stanza, except that the built code is intended for consumption by a non-Haskell client.

foreignLibName :: Lens' ForeignLib UnqualComponentName

foreignLibBuildInfo :: Lens' ForeignLib BuildInfo

foreignLibType :: Lens' ForeignLib ForeignLibType

foreignLibOptions :: Lens' ForeignLib [ForeignLibOption]

foreignLibVersionInfo :: Lens' ForeignLib (Maybe LibVersionInfo)

foreignLibVersionLinux :: Lens' ForeignLib (Maybe Version)

foreignLibModDefFile :: Lens' ForeignLib [FilePath]