Distribution.Types.GenericPackageDescription.Lens

Documentation

data GenericPackageDescription

data PackageFlag

A flag can represent a feature to be included, or a way of linking a target against its dependencies, or in fact whatever you can think of.

data FlagName

A FlagName is the name of a user-defined configuration flag

Use mkFlagName and unFlagName to convert from/to a String.

This type is opaque since Cabal-2.0

Since: Cabal-syntax-2.0.0.2

data ConfVar

A ConfVar represents the variable type used.

Constructors

OS OS

Arch Arch

PackageFlag FlagName

Impl CompilerFlavor VersionRange

packageDescription :: Lens' GenericPackageDescription PackageDescription

flagName :: Lens' PackageFlag FlagName

flagDescription :: Lens' PackageFlag String

flagDefault :: Lens' PackageFlag Bool

flagManual :: Lens' PackageFlag Bool

gpdScannedVersion :: Lens' GenericPackageDescription (Maybe Version)

genPackageFlags :: Lens' GenericPackageDescription [PackageFlag]

condLibrary :: Lens' GenericPackageDescription (Maybe (CondTree ConfVar [Dependency] Library))

condSubLibraries :: Lens' GenericPackageDescription [(UnqualComponentName, CondTree ConfVar [Dependency] Library)]

condForeignLibs :: Lens' GenericPackageDescription [(UnqualComponentName, CondTree ConfVar [Dependency] ForeignLib)]

condExecutables :: Lens' GenericPackageDescription [(UnqualComponentName, CondTree ConfVar [Dependency] Executable)]

condTestSuites :: Lens' GenericPackageDescription [(UnqualComponentName, CondTree ConfVar [Dependency] TestSuite)]

condBenchmarks :: Lens' GenericPackageDescription [(UnqualComponentName, CondTree ConfVar [Dependency] Benchmark)]

allCondTrees :: Applicative f => (forall a. CondTree ConfVar [Dependency] a -> f (CondTree ConfVar [Dependency] a)) -> GenericPackageDescription -> f GenericPackageDescription

_OS :: Traversal' ConfVar OS

_Arch :: Traversal' ConfVar Arch

_PackageFlag :: Traversal' ConfVar FlagName

_Impl :: Traversal' ConfVar (CompilerFlavor, VersionRange)