Distribution.Types.InstalledPackageInfo

Documentation

data InstalledPackageInfo

Constructors

InstalledPackageInfo

Fields

sourcePackageId :: PackageId

sourceLibName :: LibraryName

installedComponentId_ :: ComponentId

libVisibility :: LibraryVisibility

installedUnitId :: UnitId

instantiatedWith :: [(ModuleName, OpenModule)]

compatPackageKey :: String

license :: Either License License

copyright :: !ShortText

maintainer :: !ShortText

author :: !ShortText

stability :: !ShortText

homepage :: !ShortText

pkgUrl :: !ShortText

synopsis :: !ShortText

description :: !ShortText

category :: !ShortText

abiHash :: AbiHash

indefinite :: Bool

exposed :: Bool

exposedModules :: [ExposedModule]

hiddenModules :: [ModuleName]

trusted :: Bool

importDirs :: [FilePath]

libraryDirs :: [FilePath]

libraryDirsStatic :: [FilePath]

libraryDynDirs :: [FilePath]

overrides libraryDirs

dataDir :: FilePath

hsLibraries :: [String]

extraLibraries :: [String]

extraLibrariesStatic :: [String]

extraGHCiLibraries :: [String]

includeDirs :: [FilePath]

includes :: [String]

depends :: [UnitId]

abiDepends :: [AbiDependency]

ccOptions :: [String]

cxxOptions :: [String]

ldOptions :: [String]

frameworkDirs :: [FilePath]

frameworks :: [String]

haddockInterfaces :: [FilePath]

haddockHTMLs :: [FilePath]

pkgRoot :: Maybe FilePath

emptyInstalledPackageInfo :: InstalledPackageInfo

mungedPackageId :: InstalledPackageInfo -> MungedPackageId

mungedPackageName :: InstalledPackageInfo -> MungedPackageName

Returns the munged package name, which we write into name for compatibility with old versions of GHC.

data AbiDependency

An ABI dependency is a dependency on a library which also records the ABI hash (abiHash) of the library it depends on.

The primary utility of this is to enable an extra sanity when GHC loads libraries: it can check if the dependency has a matching ABI and if not, refuse to load this library. This information is critical if we are shadowing libraries; differences in the ABI hash let us know what packages get shadowed by the new version of a package.

Constructors

AbiDependency

Fields

depUnitId :: UnitId

depAbiHash :: AbiHash

data ExposedModule

Constructors

ExposedModule

Fields

exposedName :: ModuleName

exposedReexport :: Maybe OpenModule