Distribution.Types.Library

Documentation

data Library

Constructors

Library

Fields

libName :: LibraryName

exposedModules :: [ModuleName]

reexportedModules :: [ModuleReexport]

signatures :: [ModuleName]

What sigs need implementations?

libExposed :: Bool

Is the lib to be exposed by default? (i.e. whether its modules available in GHCi for example)

libVisibility :: LibraryVisibility

Whether this multilib can be used as a dependency for other packages.

libBuildInfo :: BuildInfo

emptyLibrary :: Library

explicitLibModules :: Library -> [ModuleName]

Get all the module names from the library (exposed and internal modules) which are explicitly listed in the package description which would need to be compiled. (This does not include reexports, which do not need to be compiled.) This may not include all modules for which GHC generated interface files (i.e., implicit modules.)

libModulesAutogen :: Library -> [ModuleName]

Get all the auto generated module names from the library, exposed or not. This are a subset of libModules.