Distribution.Types.PackageDescription.Lens

Documentation

data PackageDescription

This data type is the internal representation of the file pkg.cabal. It contains two kinds of information about the package: information which is needed for all packages, such as the package name and version, and information which is needed for the simple build system only, such as the compiler options and library name.

componentBuildInfo :: ComponentName -> Traversal' PackageDescription BuildInfo

Since: 2.4

allLibraries :: Traversal' PackageDescription Library

Since: 3.0.0.0

package :: Lens' PackageDescription PackageIdentifier

copyright :: Lens' PackageDescription ShortText

maintainer :: Lens' PackageDescription ShortText

author :: Lens' PackageDescription ShortText

stability :: Lens' PackageDescription ShortText

homepage :: Lens' PackageDescription ShortText

pkgUrl :: Lens' PackageDescription ShortText

synopsis :: Lens' PackageDescription ShortText

description :: Lens' PackageDescription ShortText

category :: Lens' PackageDescription ShortText

dataDir :: Lens' PackageDescription (SymbolicPath Pkg (Dir DataDir))

specVersion :: Lens' PackageDescription CabalSpecVersion

licenseRaw :: Lens' PackageDescription (Either License License)

licenseFiles :: Lens' PackageDescription [RelativePath Pkg File]

testedWith :: Lens' PackageDescription [(CompilerFlavor, VersionRange)]

bugReports :: Lens' PackageDescription ShortText

sourceRepos :: Lens' PackageDescription [SourceRepo]

customFieldsPD :: Lens' PackageDescription [(String, String)]

buildTypeRaw :: Lens' PackageDescription (Maybe BuildType)

setupBuildInfo :: Lens' PackageDescription (Maybe SetupBuildInfo)

library :: Lens' PackageDescription (Maybe Library)

subLibraries :: Lens' PackageDescription [Library]

executables :: Lens' PackageDescription [Executable]

foreignLibs :: Lens' PackageDescription [ForeignLib]

testSuites :: Lens' PackageDescription [TestSuite]

benchmarks :: Lens' PackageDescription [Benchmark]

dataFiles :: Lens' PackageDescription [RelativePath DataDir File]

extraSrcFiles :: Lens' PackageDescription [RelativePath Pkg File]

extraTmpFiles :: Lens' PackageDescription [RelativePath Pkg File]

extraDocFiles :: Lens' PackageDescription [RelativePath Pkg File]

componentModules :: Monoid r => ComponentName -> Getting r PackageDescription [ModuleName]

Since: 2.4