Distribution.Types.UnqualComponentName

Documentation

data UnqualComponentName

An unqualified component name, for any kind of component.

This is distinguished from a ComponentName and ComponentId. The former also states which of a library, executable, etc the name refers too. The later uniquely identifiers a component and its closure.

Since: Cabal-syntax-2.0.0.2

unUnqualComponentName :: UnqualComponentName -> String

Convert UnqualComponentName to String

Since: Cabal-syntax-2.0.0.2

unUnqualComponentNameST :: UnqualComponentName -> ShortText

Since: Cabal-syntax-3.4.0.0

mkUnqualComponentName :: String -> UnqualComponentName

Construct a UnqualComponentName from a String

mkUnqualComponentName is the inverse to unUnqualComponentName

Note: No validations are performed to ensure that the resulting UnqualComponentName is valid

Since: Cabal-syntax-2.0.0.2

packageNameToUnqualComponentName :: PackageName -> UnqualComponentName

Converts a package name to an unqualified component name

Useful in legacy situations where a package name may refer to an internal component, if one is defined with that name.

2018-12-21: These "legacy" situations are not legacy. We can build-depends on the internal library. However Now dependency contains Set LibraryName, and we should use that.

Since: Cabal-syntax-2.0.0.2

unqualComponentNameToPackageName :: UnqualComponentName -> PackageName

Converts an unqualified component name to a package name

packageNameToUnqualComponentName is the inverse of unqualComponentNameToPackageName.

Useful in legacy situations where a package name may refer to an internal component, if one is defined with that name.

Since: Cabal-syntax-2.0.0.2