Distribution.Types.Version

Package versions

data Version

A Version represents the version of a software entity.

Instances of Eq and Ord are provided, which gives exact equality and lexicographic ordering of the version number components (i.e. 2.1 > 2.0, 1.2.3 > 1.2.2, etc.).

This type is opaque and distinct from the Version type in Data.Version since Cabal-2.0. The difference extends to the Binary instance using a different (and more compact) encoding.

Since: Cabal-syntax-2.0.0.2

mkVersion :: [Int] -> Version

Construct Version from list of version number components.

For instance, mkVersion [3,2,1] constructs a Version representing the version 3.2.1.

All version components must be non-negative. mkVersion [] currently represents the special null version; see also nullVersion.

Since: Cabal-syntax-2.0.0.2

mkVersion' :: Version -> Version

Variant of mkVersion which converts a Data.Version Version into Cabal's Version type.

Since: Cabal-syntax-2.0.0.2

versionNumbers :: Version -> [Int]

Unpack Version into list of version number components.

This is the inverse to mkVersion, so the following holds:

(versionNumbers . mkVersion) vs == vs

Since: Cabal-syntax-2.0.0.2

nullVersion :: Version

Constant representing the special null Version

The nullVersion compares (via Ord) as less than every proper Version value.

Since: Cabal-syntax-2.0.0.2

alterVersion :: ([Int] -> [Int]) -> Version -> Version

Apply function to list of version number components

alterVersion f == mkVersion . f . versionNumbers

Since: Cabal-syntax-2.0.0.2

version0 :: Version

Version 0. A lower bound of Version.

Since: Cabal-syntax-2.2

Internal

validVersion :: Version -> Bool

versionDigitParser :: CabalParsing m => m Int

An integral without leading zeroes.

Since: Cabal-syntax-3.0