Distribution.Types.VersionRange.Internal

Description

The only purpose of this module is to prevent the export of VersionRange constructors from Distribution.Types.VersionRange. To avoid creating orphan instances, a lot of related code had to be moved here too.

Documentation

data VersionRange

Constructors

ThisVersion Version

LaterVersion Version

OrLaterVersion Version

EarlierVersion Version

OrEarlierVersion Version

MajorBoundVersion Version

UnionVersionRanges VersionRange VersionRange

IntersectVersionRanges VersionRange VersionRange

anyVersion :: VersionRange

The version range -any. That is, a version range containing all versions.

withinRange v anyVersion = True

noVersion :: VersionRange

The empty version range -none, that is a version range containing no versions.

This can be constructed using any unsatisfiable version range expression, for example < 0.

withinRange v noVersion = False

thisVersion :: Version -> VersionRange

The version range == v.

withinRange v' (thisVersion v) = v' == v

notThisVersion :: Version -> VersionRange

The version range /= v.

withinRange v' (notThisVersion v) = v' /= v

laterVersion :: Version -> VersionRange

The version range > v.

withinRange v' (laterVersion v) = v' > v

earlierVersion :: Version -> VersionRange

The version range < v.

withinRange v' (earlierVersion v) = v' < v

orLaterVersion :: Version -> VersionRange

The version range >= v.

withinRange v' (orLaterVersion v) = v' >= v

orEarlierVersion :: Version -> VersionRange

The version range <= v.

withinRange v' (orEarlierVersion v) = v' <= v

unionVersionRanges :: VersionRange -> VersionRange -> VersionRange

The version range vr1 || vr2.

  withinRange v' (unionVersionRanges vr1 vr2)
= withinRange v' vr1 || withinRange v' vr2

intersectVersionRanges :: VersionRange -> VersionRange -> VersionRange

The version range vr1 && vr2.

  withinRange v' (intersectVersionRanges vr1 vr2)
= withinRange v' vr1 && withinRange v' vr2

withinVersion :: Version -> VersionRange

The version range == v.*.

For example, for version 1.2, the version range == 1.2.* is the same as >= 1.2 && < 1.3.

withinRange v' (withinVersion v) = v' >= v && v' < upper v
  where
    upper (Version lower t) = Version (init lower ++ [last lower + 1]) t

majorBoundVersion :: Version -> VersionRange

The version range ^>= v.

For example, for version 1.2.3.4, the version range ^>= 1.2.3.4 is the same as >= 1.2.3.4 && < 1.3.

Note that ^>= 1 is equivalent to >= 1 && < 1.1.

Since: Cabal-syntax-2.0.0.2

data VersionRangeF a

F-Algebra of VersionRange. See cataVersionRange.

Since: Cabal-syntax-2.2

Constructors

ThisVersionF Version
== version.

LaterVersionF Version
> version. NB: not >=

OrLaterVersionF Version
>= version.

EarlierVersionF Version
< version.

OrEarlierVersionF Version
<= version.

MajorBoundVersionF Version
^>= version, same as >= version && < MAJ(version)+1.

UnionVersionRangesF a a
||.

IntersectVersionRangesF a a
&&.

projectVersionRange :: VersionRange -> VersionRangeF VersionRange

Generic destructor for VersionRange.

Since: Cabal-syntax-2.2

embedVersionRange :: VersionRangeF VersionRange -> VersionRange

Generic constructor for VersionRange.

Since: Cabal-syntax-2.2

cataVersionRange :: (VersionRangeF a -> a) -> VersionRange -> a

Fold VersionRange.

Since: Cabal-syntax-2.2

anaVersionRange :: (a -> VersionRangeF a) -> a -> VersionRange

Unfold VersionRange.

Since: Cabal-syntax-2.2

hyloVersionRange :: (VersionRangeF VersionRange -> VersionRange) -> (VersionRange -> VersionRangeF VersionRange) -> VersionRange -> VersionRange

Refold VersionRange.

Since: Cabal-syntax-2.2

versionRangeParser :: forall m. CabalParsing m => m Int -> CabalSpecVersion -> m VersionRange

VersionRange parser parametrised by version digit parser.

versionDigitParser is used for all VersionRange.

integral is used for backward-compat pkgconfig-depends versions, PkgConfigVersionRange.

Since: Cabal-syntax-3.0

majorUpperBound :: Version -> Version

Compute next greater major version to be used as upper bound.

Example: 0.4.1 produces the version 0.5 which then can be used to construct a range >= 0.4.1 && < 0.5

Since: Cabal-syntax-2.2

wildcardUpperBound :: Version -> Version

Increment the last version number.

Example: For 1.2 this returns 1.3 so that it can be used as upper bound when resolving == 1.2.*. For 0.4.1 it returns 0.4.2.

Since: Cabal-syntax-2.2