Distribution.Utils.Path

Symbolic path endpoints

data FileOrDir

A type-level symbolic name, to an abstract file or directory (e.g. the Cabal package directory).

Constructors

File
A file (with no further information).

Dir Type
The abstract name of a directory or category of directories, e.g. the package directory or a source directory.

data AllowAbsolute

Is this symbolic path allowed to be absolute, or must it be relative?

Constructors

AllowAbsolute
The path may be absolute, or it may be relative.

OnlyRelative
The path must be relative.

Abstract directory locations

data CWD

Abstract directory: current working directory.

See Note [Symbolic paths] in Distribution.Utils.Path.

data Pkg

Abstract directory: package directory (e.g. a directory containing the .cabal file).

See Note [Symbolic paths] in Distribution.Utils.Path.

data Dist

Abstract directory: dist directory (e.g. dist-newstyle).

See Note [Symbolic paths] in Distribution.Utils.Path.

data Source

Abstract directory: source directory (a search directory for source files).

See Note [Symbolic paths] in Distribution.Utils.Path.

data Include

Abstract directory: include directory (a search directory for CPP includes like header files, e.g. with ghc -I).

See Note [Symbolic paths] in Distribution.Utils.Path.

data Lib

Abstract directory: search directory for extra libraries.

See Note [Symbolic paths] in Distribution.Utils.Path.

data Framework

Abstract directory: MacOS framework directory.

See Note [Symbolic paths] in Distribution.Utils.Path.

data Build

Abstract directory: build directory.

See Note [Symbolic paths] in Distribution.Utils.Path.

data Artifacts

Abstract directory: directory for build artifacts, such as documentation or .hie files.

See Note [Symbolic paths] in Distribution.Utils.Path.

data PkgDB

Abstract directory: package database directory.

See Note [Symbolic paths] in Distribution.Utils.Path.

data DataDir

Abstract directory: data files directory.

See Note [Symbolic paths] in Distribution.Utils.Path.

data Mix

Abstract directory: directory for HPC .mix files.

See Note [Symbolic paths] in Distribution.Utils.Path.

data Tix

Abstract directory: directory for HPC .tix files.

See Note [Symbolic paths] in Distribution.Utils.Path.

data Tmp

Abstract directory: a temporary directory.

See Note [Symbolic paths] in Distribution.Utils.Path.

data Response

Abstract directory: directory for response files.

See Note [Symbolic paths] in Distribution.Utils.Path.

Symbolic paths

type RelativePath = SymbolicPathX 'OnlyRelative

A symbolic relative path, relative to an abstract location specified by the from type parameter.

They are *symbolic*, which means we cannot perform any IO until we interpret them (using e.g. interpretSymbolicPath).

type SymbolicPath = SymbolicPathX 'AllowAbsolute

A symbolic path which is allowed to be absolute.

They are *symbolic*, which means we cannot perform any IO until we interpret them (using e.g. interpretSymbolicPath).

data SymbolicPathX (allowAbsolute :: AllowAbsolute) (from :: Type) (to :: FileOrDir)

A symbolic path, possibly relative to an abstract location specified by the from type parameter.

They are *symbolic*, which means we cannot perform any IO until we interpret them (using e.g. interpretSymbolicPath).

Symbolic path API

getSymbolicPath :: SymbolicPathX allowAbsolute from to -> FilePath

Extract the FilePath underlying a SymbolicPathX or RelativePath, without interpreting it.

Use this function e.g. to validate the underlying filepath.

When interacting with the file system, you should instead use interpretSymbolicPath or interpretSymbolicPathCWD.

See Note [Symbolic paths] in Distribution.Utils.Path.

sameDirectory :: SymbolicPathX allowAbsolute from (Dir to)

A symbolic path from a directory to itself.

makeRelativePathEx :: HasCallStack => FilePath -> RelativePath from to

Make a RelativePath, ensuring the path is not absolute, but performing no further checks.

makeSymbolicPath :: FilePath -> SymbolicPath from to

Make a SymbolicPathX, which may be relative or absolute.

unsafeMakeSymbolicPath :: FilePath -> SymbolicPathX allowAbs from to

Make a SymbolicPathX which may be relative or absolute, without performing any checks.

Avoid using this function in new code.

coerceSymbolicPath :: SymbolicPathX allowAbsolute from to1 -> SymbolicPathX allowAbsolute from to2

Change what a symbolic path is pointing to.

unsafeCoerceSymbolicPath :: SymbolicPathX allowAbsolute from1 to1 -> SymbolicPathX allowAbsolute from2 to2

Change both what a symbolic path is pointing from and pointing to.

Avoid using this in new code.

relativeSymbolicPath :: RelativePath from to -> SymbolicPath from to

Weakening: convert a relative symbolic path to a symbolic path, "forgetting" that it is relative.

symbolicPathRelative_maybe :: SymbolicPath from to -> Maybe (RelativePath from to)

Is this symbolic path a relative symbolic path?

interpretSymbolicPath :: Maybe (SymbolicPath CWD (Dir Pkg)) -> SymbolicPathX allowAbsolute Pkg to -> FilePath

Interpret a symbolic path with respect to the given directory.

Use this function before directly interacting with the file system in order to take into account a working directory argument.

NB: when invoking external programs (such as GHC), it is preferable to set the working directory of the process and use interpretSymbolicPathCWD rather than calling this function, as this function will turn relative paths into absolute paths if the working directory is an absolute path. This can degrade error messages, or worse, break the behaviour entirely (because the program might expect certain paths to be relative).

See Note [Symbolic paths] in Distribution.Utils.Path.

General filepath API

(</>) :: PathLike p q r => p -> q -> r infixr 5

Like </>, but also supporting symbolic paths.

(<.>) :: FileLike p => p -> String -> p infixr 7

Like <.>, but also supporting symbolic paths.

takeDirectorySymbolicPath :: SymbolicPathX allowAbsolute from File -> SymbolicPathX allowAbsolute from (Dir to')

Like takeDirectory, for symbolic paths.

dropExtensionsSymbolicPath :: SymbolicPathX allowAbsolute from File -> SymbolicPathX allowAbsolute from File

Like dropExtensions, for symbolic paths.

replaceExtensionSymbolicPath :: SymbolicPathX allowAbsolute from File -> String -> SymbolicPathX allowAbsolute from File

Like replaceExtension, for symbolic paths.

normaliseSymbolicPath :: SymbolicPathX allowAbsolute from to -> SymbolicPathX allowAbsolute from to

Like normalise, for symbolic paths.

Working directory handling

interpretSymbolicPathCWD :: SymbolicPathX allowAbsolute Pkg to -> FilePath

Interpret a symbolic path, **under the assumption that the working directory is the package directory**.

Use interpretSymbolicPath instead if you need to take into account a working directory argument before directly interacting with the file system.

Use this function instead of interpretSymbolicPath when invoking a child process: set the working directory of the sub-process, and use this function, e.g.:

callGhc :: Maybe (SymbolicPath CWD (Dir Pkg))
        -> SymbolicPath (Dir Pkg) File
        -> IO ()
callGhc mbWorkDir inputFile =
  runProgramInvocation $
    programInvocationCwd mbWorkDir ghcProg [interpretSymbolicPathCWD inputFile]

In this example, programInvocationCwd sets the working directory, so it is appropriate to use interpretSymbolicPathCWD to provide its arguments.

See Note [Symbolic paths] in Distribution.Utils.Path.

absoluteWorkingDir :: Maybe (SymbolicPath CWD to) -> IO FilePath

Absolute path to the current working directory.

tryMakeRelativeToWorkingDir :: Maybe (SymbolicPath CWD (Dir dir)) -> SymbolicPath dir to -> IO (SymbolicPath dir to)

Try to make a path relative to the current working directory.

NB: this function may fail to make the path relative.

Module names

moduleNameSymbolicPath :: ModuleName -> SymbolicPathX allowAbsolute Source File

Retrieve the relative symbolic path to a Haskell module.