Distribution.Utils.Structured

Description

Structurally tag binary serialisation stream. Useful when most Binary instances are Generic derived.

Say you have a data type

data Record = Record
  { _recordFields  :: HM.HashMap Text (Integer, ByteString)
  , _recordEnabled :: Bool
  }
  deriving (Eq, Show, Generic)

instance Binary Record
instance Structured Record

then you can serialise and deserialise Record values with a structure tag by simply

structuredEncode record :: ByteString
structuredDecode lbs :: IO Record

If structure of Record changes in between, deserialisation will fail early.

Technically, Structured is not related to Binary, and may be useful in other uses.

Encoding and decoding

These functions operate like binary's counterparts, but the serialised version has a structure hash in front.

structuredEncode :: forall a. (Binary a, Structured a) => a -> ByteString

Structured encode. Encode a value to using binary serialisation to a lazy ByteString. Encoding starts with 16 byte large structure hash.

structuredEncodeFile :: (Binary a, Structured a) => FilePath -> a -> IO ()

Lazily serialise a value to a file

structuredDecode :: forall a. (Binary a, Structured a) => ByteString -> a

Structured decode. Decode a value from a lazy ByteString, reconstructing the original structure. Throws pure exception on invalid inputs.

structuredDecodeOrFailIO :: (Binary a, Structured a) => ByteString -> IO (Either String a)

structuredDecodeFileOrFail :: (Binary a, Structured a) => FilePath -> IO (Either String a)

Lazily reconstruct a value previously written to a file.

Structured class

class Typeable a => Structured a where

Class of types with a known Structure.

For regular data types Structured can be derived generically.

data Record = Record { a :: Int, b :: Bool, c :: [Char] } deriving (Generic)
instance Structured Record

Since: Cabal-syntax-3.2.0.0

Minimal complete definition

Nothing

Methods

structure :: Proxy a -> Structure

default structure :: (Generic a, GStructured (Rep a)) => Proxy a -> Structure

type MD5 = Fingerprint

structureHash :: forall a. Structured a => Proxy a -> MD5

Semantically hashStructure . structure.

structureBuilder :: Structure -> Builder

Flatten Structure into something we can calculate hash of.

As Structure can be potentially infinite. For mutually recursive types, we keep track of TypeReps, and put just TypeRep name when it's occurred another time.

genericStructure :: forall a. (Typeable a, Generic a, GStructured (Rep a)) => Proxy a -> Structure

Derive structure generically.

class GStructured (f :: Type -> Type)

Used to implement genericStructure.

Minimal complete definition

gstructured

nominalStructure :: Typeable a => Proxy a -> Structure

Use Typeable to infer name

containerStructure :: forall f a. (Typeable f, Structured a) => Proxy (f a) -> Structure

Structure type

data Structure

Structure of a datatype.

It can be infinite, as far as TypeReps involved are finite. (e.g. polymorphic recursion might cause troubles).

Constructors

Nominal !TypeRep !TypeVersion TypeName [Structure]
nominal, yet can be parametrised by other structures.

Newtype !TypeRep !TypeVersion TypeName Structure
a newtype wrapper

Structure !TypeRep !TypeVersion TypeName SopStructure
sum-of-products structure

data Tag a

Constructors

Tag

type TypeName = String

type ConstructorName = String

type TypeVersion = Word32

A semantic version of a data type. Usually 0.

type SopStructure = [(ConstructorName, [Structure])]

hashStructure :: Structure -> MD5

A MD5 hash digest of Structure.

typeVersion :: Functor f => (TypeVersion -> f TypeVersion) -> Structure -> f Structure

A van-Laarhoven lens into TypeVersion of Structure

typeVersion :: Lens' Structure TypeVersion

typeName :: Functor f => (TypeName -> f TypeName) -> Structure -> f Structure

A van-Laarhoven lens into TypeName of Structure

typeName :: Lens' Structure TypeName