Distribution.Version

Description

Exports the Version type along with a parser and pretty printer. A version is something like "1.3.3". It also defines the VersionRange data types. Version ranges are like ">= 1.2 && < 2".

Package versions

data Version

A Version represents the version of a software entity.

Instances of Eq and Ord are provided, which gives exact equality and lexicographic ordering of the version number components (i.e. 2.1 > 2.0, 1.2.3 > 1.2.2, etc.).

This type is opaque and distinct from the Version type in Data.Version since Cabal-2.0. The difference extends to the Binary instance using a different (and more compact) encoding.

Since: Cabal-syntax-2.0.0.2

version0 :: Version

Version 0. A lower bound of Version.

Since: Cabal-syntax-2.2

mkVersion :: [Int] -> Version

Construct Version from list of version number components.

For instance, mkVersion [3,2,1] constructs a Version representing the version 3.2.1.

All version components must be non-negative. mkVersion [] currently represents the special null version; see also nullVersion.

Since: Cabal-syntax-2.0.0.2

mkVersion' :: Version -> Version

Variant of mkVersion which converts a Data.Version Version into Cabal's Version type.

Since: Cabal-syntax-2.0.0.2

versionNumbers :: Version -> [Int]

Unpack Version into list of version number components.

This is the inverse to mkVersion, so the following holds:

(versionNumbers . mkVersion) vs == vs

Since: Cabal-syntax-2.0.0.2

nullVersion :: Version

Constant representing the special null Version

The nullVersion compares (via Ord) as less than every proper Version value.

Since: Cabal-syntax-2.0.0.2

alterVersion :: ([Int] -> [Int]) -> Version -> Version

Apply function to list of version number components

alterVersion f == mkVersion . f . versionNumbers

Since: Cabal-syntax-2.0.0.2

Version ranges

data VersionRange

Constructing

anyVersion :: VersionRange

The version range -any. That is, a version range containing all versions.

withinRange v anyVersion = True

noVersion :: VersionRange

The empty version range -none, that is a version range containing no versions.

This can be constructed using any unsatisfiable version range expression, for example < 0.

withinRange v noVersion = False

thisVersion :: Version -> VersionRange

The version range == v.

withinRange v' (thisVersion v) = v' == v

notThisVersion :: Version -> VersionRange

The version range /= v.

withinRange v' (notThisVersion v) = v' /= v

laterVersion :: Version -> VersionRange

The version range > v.

withinRange v' (laterVersion v) = v' > v

earlierVersion :: Version -> VersionRange

The version range < v.

withinRange v' (earlierVersion v) = v' < v

orLaterVersion :: Version -> VersionRange

The version range >= v.

withinRange v' (orLaterVersion v) = v' >= v

orEarlierVersion :: Version -> VersionRange

The version range <= v.

withinRange v' (orEarlierVersion v) = v' <= v

unionVersionRanges :: VersionRange -> VersionRange -> VersionRange

The version range vr1 || vr2.

  withinRange v' (unionVersionRanges vr1 vr2)
= withinRange v' vr1 || withinRange v' vr2

intersectVersionRanges :: VersionRange -> VersionRange -> VersionRange

The version range vr1 && vr2.

  withinRange v' (intersectVersionRanges vr1 vr2)
= withinRange v' vr1 && withinRange v' vr2

withinVersion :: Version -> VersionRange

The version range == v.*.

For example, for version 1.2, the version range == 1.2.* is the same as >= 1.2 && < 1.3.

withinRange v' (withinVersion v) = v' >= v && v' < upper v
  where
    upper (Version lower t) = Version (init lower ++ [last lower + 1]) t

majorBoundVersion :: Version -> VersionRange

The version range ^>= v.

For example, for version 1.2.3.4, the version range ^>= 1.2.3.4 is the same as >= 1.2.3.4 && < 1.3.

Note that ^>= 1 is equivalent to >= 1 && < 1.1.

Since: Cabal-syntax-2.0.0.2

Inspection

withinRange :: Version -> VersionRange -> Bool

Does this version fall within the given range?

This is the evaluation function for the VersionRange type.

isAnyVersion :: VersionRange -> Bool

Does this VersionRange place any restriction on the Version or is it in fact equivalent to AnyVersion.

Note this is a semantic check, not simply a syntactic check. So for example the following is True (for all v).

isAnyVersion (EarlierVersion v `UnionVersionRanges` orLaterVersion v)

isNoVersion :: VersionRange -> Bool

This is the converse of isAnyVersion. It check if the version range is empty, if there is no possible version that satisfies the version range.

For example this is True (for all v):

isNoVersion (EarlierVersion v `IntersectVersionRanges` LaterVersion v)

isSpecificVersion :: VersionRange -> Maybe Version

Is this version range in fact just a specific version?

For example the version range ">= 3 && <= 3" contains only the version 3.

simplifyVersionRange :: VersionRange -> VersionRange

Simplify a VersionRange expression. For non-empty version ranges this produces a canonical form. Empty or inconsistent version ranges are left as-is because that provides more information.

If you need a canonical form use fromVersionIntervals . toVersionIntervals

It satisfies the following properties:

withinRange v (simplifyVersionRange r) = withinRange v r

    withinRange v r = withinRange v r'
==> simplifyVersionRange r = simplifyVersionRange r'
 || isNoVersion r
 || isNoVersion r'

foldVersionRange

Arguments

:: a
"-any" version

-> (Version -> a)
"== v"

-> (Version -> a)
"> v"

-> (Version -> a)
"< v"

-> (a -> a -> a)
"_ || _" union

-> (a -> a -> a)
"_ && _" intersection

-> VersionRange

-> a

Fold over the basic syntactic structure of a VersionRange.

This provides a syntactic view of the expression defining the version range. The syntactic sugar ">= v", "<= v" and "== v.*" is presented in terms of the other basic syntax.

For a semantic view use asVersionIntervals.

normaliseVersionRange :: VersionRange -> VersionRange

Normalise VersionRange.

In particular collapse (== v || > v) into >= v, and so on.

stripParensVersionRange :: VersionRange -> VersionRange

Remove VersionRangeParens constructors.

Since version 3.4 this function is id, there aren't VersionRangeParens constructor in VersionRange anymore.

Since: Cabal-syntax-2.2

hasUpperBound :: VersionRange -> Bool

Does the version range have an upper bound?

Since: Cabal-syntax-1.24.0.0

hasLowerBound :: VersionRange -> Bool

Does the version range have an explicit lower bound?

Note: this function only considers the user-specified lower bounds, but not the implicit >=0 lower bound.

Since: Cabal-syntax-1.24.0.0

Cata & ana

data VersionRangeF a

F-Algebra of VersionRange. See cataVersionRange.

Since: Cabal-syntax-2.2

Constructors

ThisVersionF Version
== version.

LaterVersionF Version
> version. NB: not >=

OrLaterVersionF Version
>= version.

EarlierVersionF Version
< version.

OrEarlierVersionF Version
<= version.

MajorBoundVersionF Version
^>= version, same as >= version && < MAJ(version)+1.

UnionVersionRangesF a a
||.

IntersectVersionRangesF a a
&&.

cataVersionRange :: (VersionRangeF a -> a) -> VersionRange -> a

Fold VersionRange.

Since: Cabal-syntax-2.2

anaVersionRange :: (a -> VersionRangeF a) -> a -> VersionRange

Unfold VersionRange.

Since: Cabal-syntax-2.2

hyloVersionRange :: (VersionRangeF VersionRange -> VersionRange) -> (VersionRange -> VersionRangeF VersionRange) -> VersionRange -> VersionRange

Refold VersionRange.

Since: Cabal-syntax-2.2

projectVersionRange :: VersionRange -> VersionRangeF VersionRange

Generic destructor for VersionRange.

Since: Cabal-syntax-2.2

embedVersionRange :: VersionRangeF VersionRange -> VersionRange

Generic constructor for VersionRange.

Since: Cabal-syntax-2.2

Utilities

wildcardUpperBound :: Version -> Version

Increment the last version number.

Example: For 1.2 this returns 1.3 so that it can be used as upper bound when resolving == 1.2.*. For 0.4.1 it returns 0.4.2.

Since: Cabal-syntax-2.2

majorUpperBound :: Version -> Version

Compute next greater major version to be used as upper bound.

Example: 0.4.1 produces the version 0.5 which then can be used to construct a range >= 0.4.1 && < 0.5

Since: Cabal-syntax-2.2

Modification

removeUpperBound :: VersionRange -> VersionRange

Given a version range, remove the highest upper bound. Example: (>= 1 && < 3) || (>= 4 && < 5) is converted to (>= 1 && || (= 4).

removeLowerBound :: VersionRange -> VersionRange

Given a version range, remove the lowest lower bound. Example: (>= 1 && || (= 4 && < 5) is converted to (>= 0 && || (= 4 && < 5).

transformCaret :: VersionRange -> VersionRange

Rewrite ^>= x.y.z into >= x.y.z && < x.(y+1)

Since: Cabal-syntax-3.6.0.0

transformCaretUpper :: VersionRange -> VersionRange

Rewrite ^>= x.y.z into >= x.y.z

Since: Cabal-syntax-3.6.0.0

transformCaretLower :: VersionRange -> VersionRange

Rewrite ^>= x.y.z into <x.(y+1)

Since: Cabal-syntax-3.6.0.0

Version intervals view

asVersionIntervals :: VersionRange -> [VersionInterval]

View a VersionRange as a union of intervals.

This provides a canonical view of the semantics of a VersionRange as opposed to the syntax of the expression used to define it. For the syntactic view use foldVersionRange.

Each interval is non-empty. The sequence is in increasing order and no intervals overlap or touch. Therefore only the first and last can be unbounded. The sequence can be empty if the range is empty (e.g. a range expression like && 2).

Other checks are trivial to implement using this view. For example:

isNoVersion vr | [] <- asVersionIntervals vr = True
               | otherwise                   = False

isSpecificVersion vr
   | [(LowerBound v  InclusiveBound
      ,UpperBound v' InclusiveBound)] <- asVersionIntervals vr
   , v == v'   = Just v
   | otherwise = Nothing

data VersionInterval

Constructors

VersionInterval !LowerBound !UpperBound

data LowerBound

Constructors

LowerBound !Version !Bound

data UpperBound

Constructors

NoUpperBound

UpperBound !Version !Bound

data Bound

Constructors

ExclusiveBound

InclusiveBound

VersionIntervals abstract type

The VersionIntervals type and the accompanying functions are exposed primarily for completeness and testing purposes. In practice asVersionIntervals is the main function to use to view a VersionRange as a bunch of VersionIntervals.

data VersionIntervals

A complementary representation of a VersionRange. Instead of a boolean version predicate it uses an increasing sequence of non-overlapping, non-empty intervals.

The key point is that this representation gives a canonical representation for the semantics of VersionRanges. This makes it easier to check things like whether a version range is empty, covers all versions, or requires a certain minimum or maximum version. It also makes it easy to check equality or containment. It also makes it easier to identify 'simple' version predicates for translation into foreign packaging systems that do not support complex version range expressions.

toVersionIntervals :: VersionRange -> VersionIntervals

Convert a VersionRange to a sequence of version intervals.

fromVersionIntervals :: VersionIntervals -> VersionRange

Convert a VersionIntervals value back into a VersionRange expression representing the version intervals.

unVersionIntervals :: VersionIntervals -> [VersionInterval]

Inspect the list of version intervals.