Distribution.Parsec.Warning

Documentation

data PWarning

Parser warning.

Constructors

PWarning !PWarnType !Position String

data PWarnType

Type of parser warning. We do classify warnings.

Different application may decide not to show some, or have fatal behaviour on others

Constructors

PWTOther
Unclassified warning

PWTUTF
Invalid UTF encoding

PWTBoolCase
true or false, not True or False

PWTVersionTag
there are version with tags

PWTNewSyntax
New syntax used, but no cabal-version: >= 1.2 specified

PWTOldSyntax
Old syntax used, and cabal-version >= 1.2 specified

PWTDeprecatedField

PWTInvalidSubsection

PWTUnknownField

PWTUnknownSection

PWTTrailingFields

PWTExtraMainIs
extra main-is field

PWTExtraTestModule
extra test-module field

PWTExtraBenchmarkModule
extra benchmark-module field

PWTLexNBSP

PWTLexBOM

PWTLexTab

PWTQuirkyCabalFile
legacy cabal file that we know how to patch

PWTDoubleDash
Double dash token, most likely it's a mistake - it's not a comment

PWTMultipleSingularField
e.g. name or version should be specified only once.

PWTBuildTypeDefault
Workaround for derive-package having build-type: Default. See https://github.com/haskell/cabal/issues/5020.

PWTVersionOperator
Version operators used (without cabal-version: 1.8)

PWTVersionWildcard
Version wildcard used (without cabal-version: 1.6)

PWTSpecVersion
Warnings about cabal-version format.

PWTEmptyFilePath
Empty filepath, i.e. literally ""

PWTExperimental
Experimental feature

showPWarning :: FilePath -> PWarning -> String